fix: fix positioning
Fix positioning by removing the `absolute` position of the outer div of the `PulsatingCircle` which caused the circles to align to the next relative div from where it is implemented.

The outer div of the `PulsatingCircle` is now relative to contain the circles inside it and to keep the correct dimensions of the whole component.

BREAKING CHANGE: positioning and size changed

Positioning has been changed from "absolute" to "relative", which is why spacing may have to be adjusted.

The size of the pulse slightly changed from `250%` to `200%`.